Corrected EventDetailsViewModel. Refactored BuyTicket, JoinEventScreen and JoinEventUITest.kt  to SelectTicket[Screen/UiTest], added placeholder image for eventDetails
