+++ app/src/main/java/com/github/se/eventradar/viewmodel/EventDetailsViewModel.kt
@@ -60,7 +60,7 @@ constructor(
   }
 
   fun isTicketFree(): Boolean {
-    return !(uiState.value.ticket.price > 0.0)
+    return !(_uiState.value.ticket.price > 0.0)
   }
 }
